ExecutionAPI/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
boto3==1.28.62
python-dotenv==1.0.0
fastapi-events==0.12.2
mangum==0.17.0
aiofiles==22.1.0
python-jose[cryptography]==3.3.0
requests==2.31.0
email-validator==2.0.0